Renaming and Organization of RL algorithms in preparation for Development (#83)
I apologize in advance since this shuffling/renaming will require a bit
of updating of yamls and dependencies. My main goal is to make the
organization of the code a bit more semantically meaningful and
hopefully clearer to onboard. Here is a summary of the main changes

* Removed all algorithm specific references like DPO and PPO when we are
using these as general purpose abstractions of families of algorithms
* Moved algorithm code to an `algorithms` directory
* shuffled some files to more meaningful directories. I.e. `buffers.py`
=> `data` rather than keep only in `online`
* unified between offline, online, and reward_modeling directories to
use `model_methods.py` to implement `forward` and `loss` variants.
* Split `online_rl_loss` to `policy_loss` and `critic_loss`
* Added `OnPolicyEnum` and `Algorithm_Type` Enums for `loss_type`
similar to `PairwiseOfflineEnum`
* Made `online_rl_loss` just return `return_dict` similar to other
algorithm pipelines


MCLI run names for each pipeline:

Reward Training: `reward-reorg-aWY7m3`
Offline Training: `rebel-reorg-nboVUO`
Online Training: `grpo-reorg-fK2HDm`

[MLFlow
Link](https://dbc-559ffd80-2bfc.cloud.databricks.com/ml/experiments/2314933907661016/runs?o=7395834863327820&searchFilter=&orderByKey=attributes.start_time&orderByAsc=false&startTime=ALL&lifecycleFilter=Active&modelVersionFilter=All+Runs&datasetsFilter=W10%3D&compareRunsMode=CHART)
