+++ compose_rl/algorithms/__init__.py
@@ -1,0 +1,12 @@
+# Copyright 2024 MosaicML ComposeRL authors
+# SPDX-License-Identifier: Apache-2.0
+
+import compose_rl.algorithms.offline as offline
+import compose_rl.algorithms.online as online
+import compose_rl.algorithms.reward_modeling as reward_modeling
+
+__all__ = [
+    'offline',
+    'online',
+    'reward_modeling',
+]
