+++ compose_rl/algorithms/offline/__init__.py
@@ -1,0 +1,14 @@
+# Copyright 2024 MosaicML ComposeRL authors
+# SPDX-License-Identifier: Apache-2.0
+
+from compose_rl.algorithms.offline.callback import ReferencePolicyCallback
+from compose_rl.algorithms.offline.model import (
+    ComposerHFPairwiseOfflinePolicyLM,
+    ComposerMPTPairwiseOfflinePolicyLM,
+)
+
+__all__ = [
+    'ComposerMPTPairwiseOfflinePolicyLM',
+    'ComposerHFPairwiseOfflinePolicyLM',
+    'ReferencePolicyCallback',
+]
