+++ compose_rl/algorithms/online/generation_utils/__init__.py
@@ -1,0 +1,20 @@
+# Copyright 2024 MosaicML ComposeRL authors
+# SPDX-License-Identifier: Apache-2.0
+
+from compose_rl.algorithms.online.generation_utils.generation_utils import (
+    hf_generate,
+    vllm_generate,
+)
+from compose_rl.algorithms.online.generation_utils.vllm_utils import (
+    broadcast_to_vllm,
+    create_vllm_engines,
+    init_process_group,
+)
+
+__all__ = [
+    'broadcast_to_vllm',
+    'create_vllm_engines',
+    'init_process_group',
+    'hf_generate',
+    'vllm_generate',
+]
