+++ compose_rl/algorithms/online/model.py
@@ -1,7 +1,7 @@
 # Copyright 2024 MosaicML ComposeRL authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""PPO Composer Policy implementations."""
+"""On-Policy Online RL Composer Model implementations."""
 
 import collections
 import logging
@@ -16,13 +16,14 @@ from transformers import (
     PreTrainedTokenizerFast,
 )
 
-from compose_rl.ppo.modeling_hf import ComposerHFPolicy
-from compose_rl.ppo.modeling_mpt import MPTForPolicy
-from compose_rl.ppo.modeling_utils import (
+from compose_rl.algorithms.online.model_methods import (
+    OnPolicyEnum,
     composer_online_rl_forward,
     online_rl_loss,
 )
-from compose_rl.ppo.policy_configuration import MPTPolicyConfig
+from compose_rl.algorithms.online.modeling_hf import ComposerHFPolicy
+from compose_rl.algorithms.online.modeling_mpt import MPTForPolicy
+from compose_rl.algorithms.online.policy_configuration import MPTPolicyConfig
 from compose_rl.utils import (
     clear_mb_load_balancing_loss,
     get_mb_load_balancing_loss,
@@ -33,7 +34,7 @@ Tokenizer = Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
 log = logging.getLogger(__name__)
 
 
-class ComposerMosaicPolicy(HuggingFaceModel):
+class ComposerMPTPolicyLM(HuggingFaceModel):
 
     def __init__(
         self,
@@ -75,7 +76,11 @@ class ComposerMosaicPolicy(HuggingFaceModel):
             self.model.transformer,  # type: ignore
         )
 
-        ret_val = composer_online_rl_forward(batch, self.model)
+        ret_val = composer_online_rl_forward(
+            batch,
+            self.model,
+            OnPolicyEnum.PPO,
+        )
 
         lbl = get_mb_load_balancing_loss(
             self.config,
@@ -92,10 +97,10 @@ class ComposerMosaicPolicy(HuggingFaceModel):
         )
 
     def loss(self, outputs: MutableMapping, batch: MutableMapping):
-        return_dict, kl_loss = online_rl_loss(
+        return_dict = online_rl_loss(
             outputs=outputs,
             batch=batch,
-            loss_type='ppo',
+            loss_type=OnPolicyEnum.PPO,
             value_clip_range=self.config.value_clip_range,
             value_loss_weight=self.config.value_loss_weight,
             policy_clip_ratio=self.config.policy_clip_ratio,
@@ -104,7 +109,7 @@ class ComposerMosaicPolicy(HuggingFaceModel):
             kl_clip_range=self.config.kl_clip_range,
         )
 
-        self.policy_kl.append(kl_loss)
+        self.policy_kl.append(return_dict['kl/policy_kl'])
 
         return return_dict
 
@@ -125,7 +130,7 @@ class ComposerMosaicPolicy(HuggingFaceModel):
         self.batch_stats = batch_stats  # pyright: ignore
 
 
-class ComposerHFPolicyModel(ComposerHFPolicy):
+class ComposerHFPolicyLM(ComposerHFPolicy):
 
     def __init__(
         self,
@@ -154,7 +159,7 @@ class ComposerHFPolicyModel(ComposerHFPolicy):
             self.compute_kl_loss = config_overrides.get('compute_kl_loss')
             self.target_kl = config_overrides.get('target_kl')
 
-        self.loss_type = loss_type
+        self.loss_type = OnPolicyEnum(loss_type)
 
         # Validating the input types
         assert isinstance(self.compute_kl_loss, bool)
@@ -205,7 +210,7 @@ class ComposerHFPolicyModel(ComposerHFPolicy):
         )
 
     def loss(self, outputs: MutableMapping, batch: MutableMapping):
-        return_dict, kl_loss = online_rl_loss(
+        return_dict = online_rl_loss(
             outputs=outputs,
             batch=batch,
             loss_type=self.loss_type,  # pyright: ignore
@@ -217,7 +222,7 @@ class ComposerHFPolicyModel(ComposerHFPolicy):
             kl_clip_range=self.config.kl_clip_range,
         )
 
-        self.policy_kl.append(kl_loss)
+        self.policy_kl.append(return_dict['kl/policy_kl'])
 
         return return_dict
 
@@ -238,7 +243,7 @@ class ComposerHFPolicyModel(ComposerHFPolicy):
         self.batch_stats = batch_stats
 
 
-class ComposerHFCriticFreePolicyModel(ComposerHFCausalLM):
+class ComposerHFCriticFreePolicyLM(ComposerHFCausalLM):
     """HF class wrapper for Critic Free Policy model."""
     default_train_metrics: tuple = ()
     default_eval_metrics: tuple = ()
@@ -271,7 +276,7 @@ class ComposerHFCriticFreePolicyModel(ComposerHFCausalLM):
         """
         super().__init__(**kwargs)
         self.policy_kl = []
-        self.loss_type = loss_type
+        self.loss_type = OnPolicyEnum(loss_type)
         self.normalize_advantage = normalize_advantage
         self.length_normalize_policy_loss = length_normalize_policy_loss
         self.policy_clip_ratio = policy_clip_ratio
@@ -295,7 +300,7 @@ class ComposerHFCriticFreePolicyModel(ComposerHFCausalLM):
         )
 
     def loss(self, outputs: MutableMapping, batch: MutableMapping):
-        return_dict, kl_loss = online_rl_loss(
+        return_dict = online_rl_loss(
             outputs=outputs,
             batch=batch,
             loss_type=self.loss_type,
@@ -307,7 +312,7 @@ class ComposerHFCriticFreePolicyModel(ComposerHFCausalLM):
             kl_clip_range=self.kl_clip_range,
         )
 
-        self.policy_kl.append(kl_loss)
+        self.policy_kl.append(return_dict['kl/policy_kl'])
         return return_dict
 
     def determine_early_stop(self):
