+++ tests/functional_rewards/test_bad_generation.py
@@ -9,7 +9,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
-from compose_rl.reward_learning import BadGenerationEndReward
+from compose_rl.algorithms.reward_modeling import BadGenerationEndReward
 
 
 @pytest.fixture
