+++ tests/functional_rewards/test_increasing_numbers.py
@@ -9,7 +9,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
-from compose_rl.reward_learning import IncreasingNumbersReward
+from compose_rl.algorithms.reward_modeling import IncreasingNumbersReward
 
 
 @pytest.fixture
