+++ tests/functional_rewards/test_math_verifier_reward.py
@@ -9,7 +9,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
-from compose_rl.reward_learning import MATHVerifierReward
+from compose_rl.algorithms.reward_modeling import MATHVerifierReward
 
 
 @pytest.fixture
