+++ tests/functional_rewards/test_output_length_reward.py
@@ -7,7 +7,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
-from compose_rl.reward_learning import OutputLengthReward
+from compose_rl.algorithms.reward_modeling import OutputLengthReward
 
 
 @pytest.fixture
