+++ tests/functional_rewards/test_short_response_reward.py
@@ -9,7 +9,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer
 
-from compose_rl.reward_learning import ShortResponseReward
+from compose_rl.algorithms.reward_modeling import ShortResponseReward
 
 
 @pytest.fixture
