+++ tests/test_reward_modeling.py
@@ -24,11 +24,12 @@ from torch.utils.data import DataLoader
 from transformers import AutoTokenizer
 from transformers.models.llama.modeling_llama import LlamaAttention
 
+from compose_rl.algorithms.reward_modeling.hf_utils import \
+    AutoModelForCausalLMWithRM
 from compose_rl.data import (
     finegrained_preference_dataset_collate_fn,
     pairwise_preference_dataset_collate_fn,
 )
-from compose_rl.reward_learning.hf_utils import AutoModelForCausalLMWithRM
 from tests.common import FineGrainedPreference, PairwisePreference, world_size
